requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
idna==3.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
urllib3==2.4.0
